Add traces versioning Minor improvements
