+++ parser/V23/parserImpl_V23.go
@@ -18,7 +18,7 @@ import (
 )
 
 const (
-	Version = "v23"
+	Version = "v1.23"
 )
 
 type Parser struct {
@@ -82,7 +82,7 @@ func (p *Parser) ParseTransactions(traces []byte, tipset *types.ExtendedTipSet, 
 
 		// Fees
 		if trace.GasCost.TotalCost.Uint64() > 0 {
-			feeTx := p.feesTransactions(trace, tipset, transaction.TxType)
+			feeTx := p.feesTransactions(trace, tipset, transaction.TxType, transaction.Id)
 			transactions = append(transactions, feeTx)
 		}
 	}
@@ -217,7 +217,7 @@ func (p *Parser) parseTrace(trace typesv23.ExecutionTraceV23, msgCid cid.Cid, ti
 	}, nil
 }
 
-func (p *Parser) feesTransactions(msg *typesv23.InvocResultV23, tipset *types.ExtendedTipSet, txType string) *types.Transaction {
+func (p *Parser) feesTransactions(msg *typesv23.InvocResultV23, tipset *types.ExtendedTipSet, txType, parentTxId string) *types.Transaction {
 	timestamp := parser.GetTimestamp(tipset.MinTimestamp())
 	blockCid, err := tools.GetBlockCidFromMsgCid(msg.MsgCid.String(), txType, nil, tipset)
 	if err != nil {
@@ -247,7 +247,6 @@ func (p *Parser) feesTransactions(msg *typesv23.InvocResultV23, tipset *types.Ex
 
 	metadata, _ := json.Marshal(feesMetadata)
 	feeID := tools.BuildFeeId(tipset.GetCidString(), blockCid, msg.MsgCid.String())
-	msgId := tools.BuildMessageId(tipset.GetCidString(), blockCid, msg.MsgCid.String())
 
 	return &types.Transaction{
 		BasicBlockData: types.BasicBlockData{
@@ -256,7 +255,7 @@ func (p *Parser) feesTransactions(msg *typesv23.InvocResultV23, tipset *types.Ex
 			BlockCid:  blockCid,
 		},
 		Id:          feeID,
-		ParentId:    msgId,
+		ParentId:    parentTxId,
 		TxTimestamp: timestamp,
 		TxCid:       msg.MsgCid.String(),
 		TxFrom:      msg.Msg.From.String(),
